Added option to fill-in missing data with reference homozygote
This automatically turns on --assume-reference and is useful if the
VCF contains multiple samples that have been merged from single
sample VCF files where the merging tool has set missing data for
sites not detected in all samples singly.

Assessing performance and accuracy of this, especially when the
assumption of reference homozygote is not correct at some rate,
has not been tested. This is just functionality. Consequence of
using the option will be assessed later and best practice
recommendations to be determined.