Added option to run only the specified chromosome(s)
Presently, the steps of inserting reference homozygote sites and
filling in missing data are still run on the entire input file
and this may be a very slow step for large inputs. Need to have
run-analysis.pl first subset to individual chromosomes, then run
this script, but that will require some restructuring.