Hacks for precisionFDA
If running on pFDA, the default reference panel, which includes the
genetic map, will not be downloaded and thus
~/rfmix-reference/hapmap-phase2-genetic-map.tsv.gz
won't be there and instead will be in /work/reference, being put there
by the pFDA app. Hacks added to run-analysis.pl to detect this and then
point rfmix to that instead.

Also detect the case when there was no chromosome that was run
successfully, and exit run-analysis.pl with non-zero exit code in
this case. Previously, empty result files were created and the
script exited cleanly as any one chromosome failing does not cause
the script to exit - it will just skip to the next one. In that
fashion it may attempt to run all chromosomes, and fail on all of
them, and produce no usable results but end up create empty final
result files and appear to have succeeded.